salary table##

+++ src/Employee-API/Employee.Domain/Entities/EmployeePayment.cs
@@ -6,7 +6,7 @@ namespace Employee.Domain.Entities
     {
         public long Id { get; set; }
         public DateTime PaymentDate { get; set; }
-        public double AmountPayed { get; set; } = 0;
+        public float AmountPayed { get; set; } = 0;
         public long EmployeeDetailId { get; set; }
         public EmployeeDetail EmployeeDetail { get; set; }
     }
